feat(latest_release): Add new script
